chore(tech): move sendEmailApplicationSubmitted into a directory
